+++ app/Http/Controllers/FileController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\File;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Validator;
 
 class FileController extends Controller
 {
@@ -23,11 +24,28 @@ class FileController extends Controller
 
     public function store(Request $request)
     {
-        $this->validateFile($request);
+        $this->authorize('admin');
+
+        $this->validateFile($request->all());
 
         $newFile = File::create($request->all());
 
         return response()->json($newFile, 201);
+    }
+
+    public function storeMultiple(Request $request)
+    {
+        $this->authorize('admin');
+
+        foreach ($request->fileArray as $key => $file) {
+
+            $this->validateFile($file);
+
+        }
+
+        $newFiles = File::insert($request->fileArray);
+
+        return response()->json($newFiles, 201);
     }
 
     public function search($key)
@@ -41,9 +59,11 @@ class FileController extends Controller
 
     public function update(Request $request, $id)
     {
+        $this->authorize('admin');
+
         $file = File::findOrFail($id);
 
-        $this->validateFile($request);
+        $this->validateFile($request->all());
 
         $file->update($request->all());
 
@@ -52,13 +72,41 @@ class FileController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('admin');
+
         File::destroy($id);
+    }
+
+    public function destroyMultiple(Request $request)
+    {
+        $this->authorize('admin');
+
+        File::destroy($request->files);
+    }
+
+    public function trashed()
+    {
+        $files = File::onlyTrashed()->get();
+
+        return response()->json($files);
     }
 
     public function restore($id)
     {
+        $this->authorize('admin');
+
         $file = File::onlyTrashed()->findOrFail($id);
         $file->restore();
+
+        return response()->json($file);
+    }
+
+    public function remove($id)
+    {
+        $this->authorize('admin');
+
+        $file = File::onlyTrashed()->findOrFail($id);
+        $file->forceDelete();
 
         return response()->json($file);
     }
@@ -79,18 +127,16 @@ class FileController extends Controller
         return response()->json($count);
     }
 
-    private function validateFile(Request $request)
+    private function validateFile($file)
     {
 
-        $this->validate($request, [
+        Validator::make($file, [
             'user_id' => 'required',
             'url' => 'required',
             'filename' => 'required|min:1|max:255',
             'description' => 'required|min:1|max:255',
             'mimetype' => 'required',
-            'category_id' => 'required|exists:categories,id',
-            'department_id' => 'required|exists:departments,id',
-        ]);
+        ])->validate();
 
     }
